[Refactor] refactoring for OP
